requirements.txt
Django==5.2.1
python-telegram-bot==22.1